Disable in <#834581735642628147>

+++ modules/antiOk.py
@@ -43,6 +43,7 @@ class ChatEvents(commands.Cog):
         await a.close()
     @commands.Cog.listener(name="on_message")
     async def shoe_is_gay(self, message: discord.Message):
+        if message.channel.id == 834581735642628147: return
         if message.author.id not in (713865980526329887, 821647232801570866): return
         if message.content == "ok": await message.delete()
     @commands.Cog.listener(name="on_message")
